feat: セクションがクリッカブルでないことでトピックと差別化する
